refactor: trocando o nome da propriedade de username para email.
