+++ src/main/java/br/com/nat/quadralivre/domain/quadra/QuadraRepository.java
@@ -5,4 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuadraRepository extends JpaRepository<Quadra, Long> {
     boolean existsByNomeAndEndereco(String nome, Endereco endereco);
+
+    boolean existsByGestorEmail(String email);
 }
