+++ src/main/java/br/com/nat/quadralivre/domain/usuario/UsuarioDadosDetalhados.java
@@ -2,13 +2,13 @@ package br.com.nat.quadralivre.domain.usuario;
 
 public record UsuarioDadosDetalhados (
     Long id,
-    String login,
+    String email,
     String nome,
     String telefone,
     String cpf,
     Role role
 ){
     public UsuarioDadosDetalhados(Usuario usuario){
-        this(usuario.getId(), usuario.getLogin(), usuario.getNome(), usuario.getTelefone(), usuario.getCpf(), usuario.getRole());
+        this(usuario.getId(), usuario.getEmail(), usuario.getNome(), usuario.getTelefone(), usuario.getCpf(), usuario.getRole());
     }
 }
