+++ src/main/java/br/com/nat/quadralivre/domain/usuario/UsuarioRepository.java
@@ -1,17 +1,10 @@
 package br.com.nat.quadralivre.domain.usuario;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Optional;
-
 public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
-    UserDetails findByLogin(String login);
-
-    @Query("SELECT u FROM Usuario u WHERE u.login LIKE :login")
-    Optional<Usuario> buscarUsuarioPorCampoLogin(String login);
-
-    boolean existsByLogin(String login);
+    UserDetails findByEmail(String email);
+    boolean existsByEmail(String email);
     boolean existsByCpf(String cpf);
 }
