+++ src/main/java/br/com/nat/quadralivre/domain/usuario/autenticacao/UsuarioAutenticacao.java
@@ -1,4 +1,4 @@
-package br.com.nat.quadralivre.domain.usuario;
+package br.com.nat.quadralivre.domain.usuario.autenticacao;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Size;
 public record UsuarioAutenticacao(
         @NotBlank
         @Email(message = "Digite um e-mail válido. Exemplo: email@email.com")
-        String login,
+        String email,
         @NotBlank
         @Size(min = 6, message = "A senha deve ter pelo menos 6 caracteres." )
         String senha
