+++ src/main/java/br/com/nat/quadralivre/domain/usuario/autenticacao/UsuarioAutenticacaoService.java
@@ -1,12 +1,11 @@
-package br.com.nat.quadralivre.domain.usuario;
+package br.com.nat.quadralivre.domain.usuario.autenticacao;
 
+import br.com.nat.quadralivre.domain.usuario.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class UsuarioAutenticacaoService implements UserDetailsService {
@@ -15,7 +14,7 @@ public class UsuarioAutenticacaoService implements UserDetailsService {
     private UsuarioRepository usuarioRepository;
 
     @Override
-    public UserDetails loadUserByUsername(String login) throws UsernameNotFoundException {
-        return this.usuarioRepository.findByLogin(login);
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return this.usuarioRepository.findByEmail(email);
     }
 }
